pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "codex-repo-tool"
version = "0.1.0"
description = "Unified tool for Codex-style agents to manage and edit GitHub repositories safely."
readme = "README.md"
authors = [{name = "Your Name", email = "you@example.com"}]
license = {text = "MIT"}
requires-python = ">=3.11"
dependencies = [
    "PyYAML>=6.0.2",
    "requests>=2.32.0",
    "typing-extensions>=4.12.2"
]

[project.scripts]
codexrt = "codex_repo_tool.cli:main"

[tool.black]
line-length = 100
target-version = ["py311"]

[tool.ruff]
line-length = 100
select = ["E", "F", "I", "N", "UP"]
ignore = []

[tool.mypy]
python_version = "3.11"
warn_unused_configs = true
warn_redundant_casts = true
warn_unused_ignores = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
no_implicit_optional = true
strict_optional = true
